requirements.txt
pymilvus==2.3.4
sentence-transformers==2.2.2
ollama==0.2.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
torch
transformers==4.36.2
